requirements.txt
openai
python-docx
ollama